requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
requests==2.24.0
dialogflow==1.0.0
git+git://github.com/MasterRoshan/iss_challenge
